requirements.txt
google
google-generativeai
python-dotenv
pillow 
requests
streamlit
PyPDF2
faiss-cpu
openai
pypdf
numpy 
pypdf
pydantic<2.0.0
langchain==0.3.25
langchain-core==0.3.61
langchain-community
langchain-google-genai
fpdf
langgraph



